backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "minio",
    "python-dotenv",
    "matplotlib",
    "pandas"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
